refactor: colocate show-page components under app/show/[id]
