+++ web/app/show/[id]/_components/elevation-box.test.tsx
@@ -3,9 +3,9 @@ import React from 'react'
 import '@testing-library/jest-dom'
 import { initialize, LatLng } from '@googlemaps/jest-mocks'
 import { Mock } from 'vitest'
-import { useConfig } from '../utils/config'
-import { useData } from '../utils/data-context'
-import { useMapContext } from '../utils/map-context'
+import { useConfig } from '@/lib/utils/config'
+import { useData } from '@/lib/utils/data-context'
+import { useMapContext } from '@/lib/utils/map-context'
 import ElevationBox from './elevation-box'
 
 // Rechartsのモック
@@ -21,15 +21,15 @@ vi.mock('recharts', () => ({
   Tooltip: () => <div data-testid="tooltip" />,
 }))
 
-vi.mock('../utils/config', () => ({
+vi.mock('@/lib/utils/config', () => ({
   useConfig: vi.fn(),
 }))
 
-vi.mock('../utils/data-context', () => ({
+vi.mock('@/lib/utils/data-context', () => ({
   useData: vi.fn(),
 }))
 
-vi.mock('../utils/map-context', () => ({
+vi.mock('@/lib/utils/map-context', () => ({
   useMapContext: vi.fn(),
 }))
