+++ web/app/show/[id]/_components/panorama-box.test.tsx
@@ -2,20 +2,20 @@ import { initialize, LatLng } from '@googlemaps/jest-mocks'
 import { fireEvent, render, screen } from '@testing-library/react'
 import React from 'react'
 import { Mock } from 'vitest'
-import { useData } from '../utils/data-context'
-import { useMainContext } from '../utils/main-context'
-import { useMapContext } from '../utils/map-context'
+import { useData } from '@/lib/utils/data-context'
+import { useMainContext } from '@/lib/utils/main-context'
+import { useMapContext } from '@/lib/utils/map-context'
 import PanoramaBox from './panorama-box'
 
-vi.mock('../utils/main-context', () => ({
+vi.mock('@/lib/utils/main-context', () => ({
   useMainContext: vi.fn(),
 }))
 
-vi.mock('../utils/data-context', () => ({
+vi.mock('@/lib/utils/data-context', () => ({
   useData: vi.fn(),
 }))
 
-vi.mock('../utils/map-context', () => ({
+vi.mock('@/lib/utils/map-context', () => ({
   useMapContext: vi.fn(),
 }))
